// compumundo-app-master/src/app/interfaces/productoCargado.ts
export interface ProductoCargado {
    IDTipoPrecio?: number;
    index?: number;
    id?: string;
    Rango3?: number;
    total?: number;
    PrecioRango3?: number;
    cantidad?: number;
    Rango2?: number;
    PrecioRango2?: number;
    Rango1?: number;
    PrecioRango1?: number;
    PrecioSugerido?: number;
    Cantidad: number
    PrecioUnitario?: number
    Subtotal: number
    UnidadxPresentacion: number
    idProducto: string;
    bloqueado?: boolean;
    IDInterno?: number;
    IDEmpresaDoc?: string;
    CODTipoMovimientoDoc?: string;
    CODBodega?: string;
    Lote?: string;
    FechaVencimiento?: string;
}